refactor realm settings
also fixes: #1038
